+++ pages/store/[id]/index.tsx
