+++ pages/store/[id]/payment/index.tsx
