+++ src/components/customer/search/SearchList.tsx
@@ -85,7 +85,7 @@ const SearchList = (props: SearchListProps) => {
         <ul>
           {props.queries.map((query) => (
             <li key={query} css={[queryList, recentQueries]}>
-              <Link css={link} href={{ pathname: "/customer/stores", query: { query } }}>
+              <Link css={link} href={{ pathname: "/stores", query: { query } }}>
                 {query}
               </Link>
               <button onClick={deleteRecentQuery}>
@@ -106,7 +106,7 @@ const SearchList = (props: SearchListProps) => {
         <ul>
           {props.queries.map((query) => (
             <li key={query} css={queryList}>
-              <Link css={link} href={{ pathname: "/customer/stores", query: { query } }}>
+              <Link css={link} href={{ pathname: "/stores", query: { query } }}>
                 {query}
               </Link>
             </li>
